feature: Desarrollo de formulario UX y tambien de reorganizacion
