+++ apps/web/components/ui/button.tsx
@@ -1,5 +1,5 @@
 import type { ButtonHTMLAttributes, ReactNode } from "react";
-import { cn } from "@/lib/utils";
+import { cn } from "@/src/lib/utils";
 
 type ButtonVariant = "primary" | "ghost";
