+++ apps/web/components/ui/card.tsx
@@ -1,5 +1,5 @@
 import type { HTMLAttributes, ReactNode } from "react";
-import { cn } from "@/lib/utils";
+import { cn } from "@/src/lib/utils";
 
 type CardSectionProps = HTMLAttributes<HTMLDivElement> & {
   children: ReactNode;
