+++ apps/web/features/auth/components/auth-context.ts
@@ -1,4 +1,1 @@
-import { createContext } from "react";
-import type { AuthContextValue } from "../types/auth.types";
-
-export const AuthContext = createContext<AuthContextValue | null>(null);
+export { AuthContext } from "../context/auth-context";
