+++ apps/web/features/auth/context/auth-context.ts
@@ -1,0 +1,4 @@
+import { createContext } from "react";
+import type { AuthContextValue } from "../types/auth.types";
+
+export const AuthContext = createContext<AuthContextValue | null>(null);
