+++ apps/web/features/auth/schemas/register.schema.ts
@@ -1,5 +1,4 @@
-import { RegisterSchema } from "@logistica/shared";
-import type { z } from "zod";
-
-export const registerSchema = RegisterSchema;
-export type RegisterSchemaValues = z.infer<typeof registerSchema>;
+export {
+  registerSchema,
+  type RegisterSchemaValues,
+} from "@/src/lib/forms/schemas/auth.schema";
