+++ apps/web/features/auth/services/access-token-store.ts
@@ -1,13 +1,5 @@
-let accessToken: string | null = null;
-
-export function getAccessToken(): string | null {
-  return accessToken;
-}
-
-export function setAccessToken(nextAccessToken: string): void {
-  accessToken = nextAccessToken;
-}
-
-export function clearAccessToken(): void {
-  accessToken = null;
-}
+export {
+  clearAccessToken,
+  getAccessToken,
+  setAccessToken,
+} from "./session/access-token-store";
