+++ apps/web/features/auth/services/session/access-token-store.ts
@@ -1,0 +1,13 @@
+let accessToken: string | null = null;
+
+export function getAccessToken(): string | null {
+  return accessToken;
+}
+
+export function setAccessToken(nextAccessToken: string | null): void {
+  accessToken = nextAccessToken;
+}
+
+export function clearAccessToken(): void {
+  accessToken = null;
+}
