+++ apps/web/lib/utils.ts
@@ -1,3 +1,1 @@
-export function cn(...classes: Array<string | false | null | undefined>): string {
-  return classes.filter(Boolean).join(" ");
-}
+export { cn } from "../src/lib/utils";
