+++ apps/web/src/lib/forms/schemas/auth.schema.ts
@@ -1,0 +1,8 @@
+import { LoginSchema, RegisterSchema } from "@logistica/shared";
+import type { z } from "zod";
+
+export const loginSchema = LoginSchema;
+export const registerSchema = RegisterSchema;
+
+export type LoginSchemaValues = z.infer<typeof loginSchema>;
+export type RegisterSchemaValues = z.infer<typeof registerSchema>;
